Re-implement Ore Excavation integration
Unfortunately not very sustainable, since the source code of Ore Excavation is no longer public.
But it seems to work for now, so I'll leave it in for now.

Because of the aggressive obfuscation on the Ore Excavation jar file, we also can't depend on it for building, so this commit also includes a fake API jar that only contains those APIs that the integration code needs to compile.